Introduce benchmark support
This commit introduces a new `bench` binary which is able to take a save
and simulate a given amount of frames. There is also the ability to dump
a hash by hashing the data inside a simulation Snapshot with xxHash.

Along with this, there is also support for the Tracy profiler.

Typical setup: `meson build -Dbuildtype=debugoptimized -Dbuild_bench=true -Denable_tracy=true -Dtracy:tracy_only_localhost=true`